Mr comments resolved. (2)
