+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient.swift
@@ -1,5 +1,4 @@
 import Combine
-import Dependencies
 import Foundation
 import Networking
 
@@ -9,18 +8,4 @@ public struct RocketsClient {
 
   public var getRocket: GetRocketFunction
   public var getAllRockets: GetAllRocketsFunction
-}
-
-enum RocketsClientKey: DependencyKey {
-  public static var liveValue: RocketsClient = .liveValue
-  #if DEBUG
-  public static var testValue: RocketsClient = .testValue
-  #endif
-}
-
-extension DependencyValues {
-  var rocketsClient: RocketsClient {
-    get { self[RocketsClientKey.self] }
-    set { self[RocketsClientKey.self] = newValue }
-  }
 }
