--- RocketApp/Infrastructure/Sources/NetworkClientExtensions/DependencyKey.swift
@@ -1,8 +1,0 @@
-import Dependencies
-import Foundation
-import Networking
-
-enum NetworkClientKey: DependencyKey {
-  public static var liveValue: NetworkClientType = NetworkClient.liveValue
-  public static var testValue: NetworkClientType = NetworkClient.testValue
-}
