+++ RocketApp/Infrastructure/Sources/NetworkClientExtensions/NetworkClient+liveValue.swift
@@ -1,0 +1,11 @@
+import Dependencies
+import Foundation
+import Networking
+
+extension NetworkClient: DependencyKey {
+  public static var liveValue = NetworkClient(
+    urlSessionConfiguration: .default,
+    urlRequester: .live,
+    networkMonitorClient: .live(onQueue: .main)
+  )
+}
