+++ RocketApp/Infrastructure/Sources/NetworkClientExtensions/NetworkClient+testValue.swift
@@ -1,15 +1,8 @@
-import Dependencies
 import Foundation
 import Networking
 import XCTestDynamicOverlay
 
-extension NetworkClient: DependencyKey {
-  public static var liveValue = NetworkClient(
-    urlSessionConfiguration: .default,
-    urlRequester: .live,
-    networkMonitorClient: .live(onQueue: .main)
-  )
-
+extension NetworkClient {
   public static var testValue = Self(
     urlSessionConfiguration: unimplemented("\(Self.self).urlSessionConfiguration"),
     urlRequester: unimplemented("\(Self.self).urlRequester"),
